cli/requirements.txt
pyyaml>=6.0
jsonschema>=4.20
rich>=13.0
jinja2>=3.1
anthropic>=0.40
openai>=1.50
requests>=2.31
mkdocs>=1.6
mkdocs-material>=9.5
